handle bytes or hex or Checksum in more places
