+++ qsum/data/tests/test_data_logic.py
@@ -3,15 +3,11 @@ import hashlib
 
 import pytest
 
-from qsum import Checksum
 from qsum.core.constants import DEFAULT_HASH_ALGO
 from qsum.core.exceptions import QSumInvalidDataTypeException, QSumInvalidChecksum
 from qsum.data import data_checksum
 from qsum.data.data_logic import data_digest_from_checksum, resolve_hash_algo
-
-CHECKSUM_OBJS = [Checksum('123'),
-                 '0001a665a45920422f9d417e4867efdc4fb8a04a1f3fff1fa07e998e86f7f7a27ae3',
-                 bytes.fromhex('0001a665a45920422f9d417e4867efdc4fb8a04a1f3fff1fa07e998e86f7f7a27ae3')]
+from qsum.tests.helpers import STR_CHECKSUM_OBJS
 
 
 class Custom:
@@ -25,7 +21,7 @@ def test_invalid_type():
     _ = data_checksum(custom, type(custom), hash_algo=DEFAULT_HASH_ALGO)
 
 
-@pytest.mark.parametrize('checksum_obj', CHECKSUM_OBJS)
+@pytest.mark.parametrize('checksum_obj', STR_CHECKSUM_OBJS)
 def test_data_digest_from_checksum(checksum_obj):
     data_digest = data_digest_from_checksum(checksum_obj)
     if not isinstance(checksum_obj, str):
