+++ qsum/types/tests/test_type_logic.py
@@ -5,7 +5,8 @@ import pytest
 
 from qsum import checksum
 from qsum.core.exceptions import QSumInvalidTypeException, QSumInvalidPrefixException
-from qsum.types.type_logic import prefix_to_type
+from qsum.tests.helpers import INT_CHECKSUM_OBJS
+from qsum.types.type_logic import prefix_to_type, checksum_to_type
 from qsum.types.type_map import RESERVED_INVALID_PREFIX
 
 
@@ -22,3 +23,8 @@ def test_invalid_type():
 @pytest.mark.xfail(raises=QSumInvalidPrefixException)
 def test_prefix_to_type():
     prefix_to_type(RESERVED_INVALID_PREFIX)
+
+
+@pytest.mark.parametrize('checksum_obj', INT_CHECKSUM_OBJS)
+def test_checksum_to_type(checksum_obj):
+    assert checksum_to_type(checksum_obj) == int
